requirements.txt
pyyaml
hist
mplhep
numpy
pandas
matplotlib
rich
scipy
uproot
